done update database

+++ src/main/java/com/vn/ivs/ctu/dao/SumarizationDAO.java
@@ -6,9 +6,9 @@ import com.vn.ivs.ctu.entity.Summarization;
 
 public interface SumarizationDAO {
 	
-	public int createOrUpdate(Summarization sum);
+	public long createOrUpdate(Summarization sum);
 	public List<Summarization> getSumByClubPreMonth(int idClub,int month,int year);
 	public List<Summarization> getSumByMemberPreMonth(long idMember, int month,int year);
-	public Summarization getSumById(int id);
+	public Summarization getSumById(long id);
 	public List<Summarization> getSumByMember(long idMember, int month, int year);
 }
